% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_bf.R
\name{apply_batch_field_transform}
\alias{apply_batch_field_transform}
\title{Apply Field Transformation}
\usage{
apply_batch_field_transform(val, transform_name)
}
\arguments{
\item{val}{Vector to transform}

\item{transform_name}{Name of transformation: "factor_to_numeric", "logical_to_int"}
}
\value{
Transformed vector
}
\description{
Applies a named transformation to a vector during batch preparation.
}
\keyword{internal}
