% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_internal.R,
%   R/optimization_internal_backup.R
\name{build_convergence_trace}
\alias{build_convergence_trace}
\title{Build Convergence Trace from Archive}
\usage{
build_convergence_trace(archive_df, opt_type, objectives)

build_convergence_trace(archive_df, opt_type, objectives)
}
\arguments{
\item{archive_df}{Archive data frame}

\item{opt_type}{Optimization type}

\item{objectives}{rctbp_objectives object}
}
\value{
Data frame with convergence trace

Data frame with convergence trace
}
\description{
Build Convergence Trace from Archive

Build Convergence Trace from Archive
}
\keyword{internal}
