% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{build_predefined_from_brmsfit_legacy}
\alias{build_predefined_from_brmsfit_legacy}
\title{Build Legacy Model from Cached brmsfit}
\usage{
build_predefined_from_brmsfit_legacy(model_name, brmsfit)
}
\arguments{
\item{model_name}{Name of the predefined model}

\item{brmsfit}{Cached brmsfit object}
}
\value{
Legacy rctbp_model object
}
\description{
Build Legacy Model from Cached brmsfit
}
\keyword{internal}
