% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_builders.R
\name{build_report.rctbp_design}
\alias{build_report.rctbp_design}
\title{Build Report for rctbp_design}
\usage{
build_report.rctbp_design(x)
}
\arguments{
\item{x}{rctbp_design object}
}
\value{
List with report sections
}
\description{
Creates structured report data for a design object.
}
\keyword{internal}
