% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_builders.R
\name{build_sim_params_section}
\alias{build_sim_params_section}
\title{Build Simulation Function Parameters Section}
\usage{
build_sim_params_section(sim_fn)
}
\arguments{
\item{sim_fn}{The simulation function from a design object}
}
\value{
A report section list, or NULL if no parameters
}
\description{
Build Simulation Function Parameters Section
}
\keyword{internal}
