% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design.R
\name{create_ancova_sim_fn}
\alias{create_ancova_sim_fn}
\title{Create ANCOVA Simulation Function}
\usage{
create_ancova_sim_fn(n_arms)
}
\arguments{
\item{n_arms}{Number of arms}
}
\value{
rctbp_sim_fn object (callable with schema)
}
\description{
Internal helper to create simulation function for ANCOVA models.
Returns an \code{rctbp_sim_fn} object with test arguments and output schema.
}
\keyword{internal}
