% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_internal.R
\name{create_codomain}
\alias{create_codomain}
\title{Create Codomain from Objectives Specification}
\usage{
create_codomain(objectives)
}
\arguments{
\item{objectives}{rctbp_objectives object}
}
\value{
paradox::ParamSet for codomain
}
\description{
Create Codomain from Objectives Specification
}
\keyword{internal}
