% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{create_legacy_ancova_sim_fn}
\alias{create_legacy_ancova_sim_fn}
\title{Create Legacy ANCOVA Simulation Function}
\usage{
create_legacy_ancova_sim_fn(n_arms)
}
\arguments{
\item{n_arms}{Number of arms}
}
\value{
Simulation function
}
\description{
Create Legacy ANCOVA Simulation Function
}
\keyword{internal}
