% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_optimize.R
\name{create_pareto_codomain}
\alias{create_pareto_codomain}
\title{Create Codomain for Pareto Optimization}
\usage{
create_pareto_codomain(objectives)
}
\arguments{
\item{objectives}{Named list with directions}
}
\value{
paradox::ParamSet
}
\description{
Create Codomain for Pareto Optimization
}
\keyword{internal}
