% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design.R
\name{create_sim_fn_for_model}
\alias{create_sim_fn_for_model}
\title{Create Simulation Function for Model}
\usage{
create_sim_fn_for_model(model_name)
}
\arguments{
\item{model_name}{Name of the predefined model}
}
\value{
Simulation function
}
\description{
Internal helper to create the simulation function for a predefined model.
}
\keyword{internal}
