% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_internal.R
\name{create_worst_result}
\alias{create_worst_result}
\title{Create Worst Result for Failed Evaluations}
\usage{
create_worst_result(obj_info)
}
\arguments{
\item{obj_info}{Parsed objective info}
}
\value{
Named list with worst values
}
\description{
Create Worst Result for Failed Evaluations
}
\keyword{internal}
