% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_internal.R
\name{derive_search_types}
\alias{derive_search_types}
\title{Derive Search Types from Specification}
\usage{
derive_search_types(search, search_types, design)
}
\arguments{
\item{search}{Named list of search bounds}

\item{search_types}{User-provided type specifications (can be partial)}

\item{design}{rctbp_design object}
}
\value{
Complete named list of search types for all parameters
}
\description{
For each search parameter, determines the appropriate transform type
either from explicit \code{search_types} or by auto-inference.
}
\keyword{internal}
