% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq_function_eic.R
\name{.build_AcqFunctionEIC}
\alias{.build_AcqFunctionEIC}
\title{Build the AcqFunctionEIC R6 class definition}
\usage{
.build_AcqFunctionEIC()
}
\value{
An R6ClassGenerator
}
\description{
Called from \code{.onLoad} when mlr3mbo is available. Separated to keep the class
body in this file rather than in zzz.R.
}
\keyword{internal}
