% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{estimate_posterior_brms}
\alias{estimate_posterior_brms}
\title{Estimate Posterior Using brms Backend}
\usage{
estimate_posterior_brms(data, brms_model, backend_args = list())
}
\arguments{
\item{data}{Data frame containing the observed data}

\item{brms_model}{Pre-compiled brmsfit template model (chains = 0)}

\item{backend_args}{Named list of brms-specific arguments (chains, iter, cores, etc.)}
}
\value{
A fitted brmsfit object containing posterior samples
}
\description{
Fits a brms model to observed data by updating a pre-compiled template model.
}
\keyword{internal}
