% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_single.R
\name{evaluate_power_at_n}
\alias{evaluate_power_at_n}
\title{Evaluate Power at a Single Sample Size}
\usage{
evaluate_power_at_n(
  design,
  n,
  constant,
  n_sims,
  n_cores,
  bf_args,
  brms_args,
  verbose
)
}
\description{
Evaluate Power at a Single Sample Size
}
\keyword{internal}
