% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_system.R
\name{export_report}
\alias{export_report}
\title{Export Report to File}
\usage{
export_report(object, file, format = "markdown")
}
\arguments{
\item{object}{An S7 object (rctbp_model, rctbp_design, rctbp_conditions, or rctbp_power_analysis)}

\item{file}{Character string: path to output file}

\item{format}{Character string: output format (currently only "markdown" supported)}
}
\value{
Invisibly returns the file path
}
\description{
Exports a report for an rctbayespower object to a markdown file.
}
\examples{
\dontrun{
design <- build_design(model_name = "ancova_cont_2arms",
                       target_params = "b_arm2",
                       p_sig_scs = 0.975, p_sig_ftl = 0.5)
export_report(design, "design_report.md")
}

}
