% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_optimize.R
\name{extract_pareto_front}
\alias{extract_pareto_front}
\title{Extract Pareto Front from Instance}
\usage{
extract_pareto_front(instance, objectives)
}
\arguments{
\item{instance}{bbotk OptimInstanceBatchMultiCrit}

\item{objectives}{Named list of objectives}
}
\value{
Data frame of Pareto-optimal solutions
}
\description{
Extract Pareto Front from Instance
}
\keyword{internal}
