% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_single.R
\name{extract_single_result}
\alias{extract_single_result}
\title{Extract Single-Objective Result from Archive}
\usage{
extract_single_result(
  archive,
  design,
  target_power,
  surrogate_fit,
  surrogate_type,
  score_config,
  n_sims
)
}
\description{
Extract Single-Objective Result from Archive
}
\keyword{internal}
