% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_optimize.R
\name{find_knee_linear}
\alias{find_knee_linear}
\title{Find Knee Point Using Linear Method}
\usage{
find_knee_linear(pareto_df, obj1, obj2, objectives)
}
\arguments{
\item{pareto_df}{Data frame of Pareto-optimal solutions}

\item{obj1}{Name of first objective}

\item{obj2}{Name of second objective}

\item{objectives}{Named list with directions}
}
\value{
Data frame with single selected row
}
\description{
Selects the point with maximum perpendicular distance from the line
connecting the endpoints of the Pareto front.
}
\keyword{internal}
