% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_optimize.R
\name{find_knee_min_cost}
\alias{find_knee_min_cost}
\title{Find Knee Point by Minimizing Second Objective}
\usage{
find_knee_min_cost(pareto_df, obj2)
}
\arguments{
\item{pareto_df}{Data frame of Pareto-optimal solutions}

\item{obj2}{Name of second objective (to minimize)}
}
\value{
Data frame with single selected row
}
\description{
Find Knee Point by Minimizing Second Objective
}
\keyword{internal}
