% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_optimize.R
\name{find_knee_utopia}
\alias{find_knee_utopia}
\title{Find Knee Point Using Utopia Distance}
\usage{
find_knee_utopia(pareto_df, obj1, obj2, objectives)
}
\arguments{
\item{pareto_df}{Data frame of Pareto-optimal solutions}

\item{obj1}{Name of first objective}

\item{obj2}{Name of second objective}

\item{objectives}{Named list with directions}
}
\value{
Data frame with single selected row
}
\description{
Selects the Pareto point closest to the ideal point in normalized space.
}
\keyword{internal}
