% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_builders.R
\name{fmt_params}
\alias{fmt_params}
\title{Format Parameters Compactly}
\usage{
fmt_params(params)
}
\arguments{
\item{params}{Named list of parameters}
}
\value{
Character string with formatted parameters
}
\description{
Formats a named list of parameters as "name=value, name=value" string.
}
\keyword{internal}
