% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_builders.R
\name{format_default_value}
\alias{format_default_value}
\title{Format a Default Value for Display}
\usage{
format_default_value(val)
}
\arguments{
\item{val}{A default value from a function formal}
}
\value{
Character string representation
}
\description{
Format a Default Value for Display
}
\keyword{internal}
