% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_mbo.R
\name{generate_initial_design}
\alias{generate_initial_design}
\title{Generate Initial Design Points}
\usage{
generate_initial_design(domain, n, method)
}
\arguments{
\item{domain}{paradox::ParamSet}

\item{n}{Number of points}

\item{method}{Design method ("lhs", "random", "sobol")}
}
\value{
data.table with initial design
}
\description{
Generate Initial Design Points
}
\keyword{internal}
