% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_single.R
\name{generate_single_bo_initial}
\alias{generate_single_bo_initial}
\title{Generate Initial Design for Single-Objective BO}
\usage{
generate_single_bo_initial(n_range, n_init)
}
\arguments{
\item{n_range}{Numeric vector c(n_min, n_max)}

\item{n_init}{Number of initial points}
}
\value{
Numeric vector of sample sizes (integer-valued)
}
\description{
Creates n_init points on log(n) scale within n_range.
}
\keyword{internal}
