% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_condition.R
\name{get_args_without_defaults}
\alias{get_args_without_defaults}
\title{Get Function Arguments Without Default Values}
\usage{
get_args_without_defaults(fn)
}
\arguments{
\item{fn}{A function object to analyze}
}
\value{
A character vector containing the names of arguments without default values
}
\description{
Extracts the names of function arguments that do not have default values.
This is a utility function used for identifying required parameters.
}
\examples{
# Define a function with mixed default/non-default arguments
test_fn <- function(a, b = 1, c, d = "default") {}
get_args_without_defaults(test_fn)  # Returns c("a", "c")

}
\keyword{internal}
