% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{get_bf_parameter_names_legacy}
\alias{get_bf_parameter_names_legacy}
\title{Extract Parameter Names from BayesFlow Model (Legacy)}
\usage{
get_bf_parameter_names_legacy(model)
}
\arguments{
\item{model}{A keras/BayesFlow model object}
}
\value{
Character vector of parameter names
}
\description{
Extract Parameter Names from BayesFlow Model (Legacy)
}
\keyword{internal}
