% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design.R
\name{get_model_builder}
\alias{get_model_builder}
\title{Get Model Builder Function}
\usage{
get_model_builder(model_name)
}
\arguments{
\item{model_name}{Name of the predefined model}
}
\value{
Builder function
}
\description{
Internal helper to get the builder function for a predefined model.
}
\keyword{internal}
