% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_system.R
\name{get_output_mode}
\alias{get_output_mode}
\title{Get Current Output Mode}
\usage{
get_output_mode()
}
\value{
Character string: "cli" or "markdown"
}
\description{
Returns the current output mode setting for the package.
}
\examples{
get_output_mode()

}
