% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimization.R
\name{get_pareto_axis_labels}
\alias{get_pareto_axis_labels}
\title{Get Axis Labels for Pareto Plot}
\usage{
get_pareto_axis_labels(opt_type, x_obj, y_obj)
}
\arguments{
\item{opt_type}{Optimization type}

\item{x_obj}{X objective name}

\item{y_obj}{Y objective name}
}
\value{
List with x and y labels
}
\description{
Get Axis Labels for Pareto Plot
}
\keyword{internal}
