% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{get_predefined_model}
\alias{get_predefined_model}
\title{Get a predefined model by name}
\usage{
get_predefined_model(model_name)
}
\arguments{
\item{model_name}{Character scalar, name of the predefined model.}
}
\value{
The model object.
}
\description{
Access an internal predefined model by name.
}
