% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity.R
\name{get_verbosity}
\alias{get_verbosity}
\title{Get Current Verbosity Level}
\usage{
get_verbosity()
}
\value{
Integer: 0 (quiet), 1 (normal), or 2 (verbose)
}
\description{
Returns the current verbosity level setting for the package.
}
\examples{
get_verbosity()

}
