% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_internal.R
\name{invlogit_transform}
\alias{invlogit_transform}
\title{Inverse Logit Transform}
\usage{
invlogit_transform(x)
}
\arguments{
\item{x}{Logit-scale value(s)}
}
\value{
Probability value(s) in (0, 1)
}
\description{
Transforms unbounded value back to probability scale.
}
\keyword{internal}
