% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_system.R
\name{markdown_helpers}
\alias{markdown_helpers}
\alias{md_h1}
\alias{md_h2}
\alias{md_h3}
\alias{md_rule}
\alias{md_alert}
\alias{md_bullet}
\alias{md_code}
\alias{md_code_block}
\title{Markdown Helper Functions}
\usage{
md_h1(text)

md_h2(text)

md_h3(text)

md_rule()

md_alert(text, type = c("info", "success", "warning", "danger"))

md_bullet(items, indent = 0)

md_code(text)

md_code_block(text, lang = "")
}
\description{
Convert semantic output elements to markdown format.
}
\keyword{internal}
