% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_system.R
\name{new_output_builder}
\alias{new_output_builder}
\title{Build Output as Character Vector}
\usage{
new_output_builder()
}
\value{
Character vector that can be collapsed and rendered
}
\description{
Helper to collect output sections before rendering.
}
\keyword{internal}
