% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{obj_max_power}
\alias{obj_max_power}
\title{Objective Specification: Maximize Power}
\usage{
obj_max_power(power_metric = "pwr_eff")
}
\arguments{
\item{power_metric}{Which power metric: "pwr_eff" or "pwr_fut"}
}
\value{
List with \code{objectives} for \code{\link[=build_objectives]{build_objectives()}}
}
\description{
Creates objective specification for maximizing power.
}
\examples{
\dontrun{
spec <- obj_max_power()
obj <- build_objectives(
  design = design,
  search = list(n_total = c(50, 500)),
  objectives = spec$objectives,
  constant = list(...)
)
}
}
