% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{parse_objectives}
\alias{parse_objectives}
\title{Parse Objectives Specification}
\usage{
parse_objectives(objectives)
}
\arguments{
\item{objectives}{Named list of objectives}
}
\value{
List with parsed information
}
\description{
Parse Objectives Specification
}
\keyword{internal}
