% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_optimize.R
\name{parse_simplex_specs}
\alias{parse_simplex_specs}
\title{Parse Simplex Search Specifications}
\usage{
parse_simplex_specs(search, design)
}
\arguments{
\item{search}{Named list of search bounds}

\item{design}{rctbp_design object}
}
\value{
Named list of parsed simplex specs
}
\description{
Parse Simplex Search Specifications
}
\keyword{internal}
