% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimization.R
\name{plot_optimization_convergence}
\alias{plot_optimization_convergence}
\title{Plot Optimization Convergence}
\usage{
plot_optimization_convergence(result, show_all = TRUE, ...)
}
\arguments{
\item{result}{rctbp_optimization_result object}

\item{show_all}{Logical, show all evaluations or just cumulative best}

\item{...}{Additional ggplot arguments}
}
\value{
ggplot2 object
}
\description{
Shows how the objective value improves over evaluations.
}
