% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimization.R
\name{plot_optimization_pareto}
\alias{plot_optimization_pareto}
\title{Plot Pareto Front}
\usage{
plot_optimization_pareto(result, highlight_best = TRUE, ...)
}
\arguments{
\item{result}{rctbp_optimization_result object}

\item{highlight_best}{Logical, highlight the "knee" point}

\item{...}{Additional ggplot arguments}
}
\value{
ggplot2 object
}
\description{
Visualizes the Pareto-optimal solutions for multi-objective optimization.
}
