% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimization.R
\name{plot_optimization_search}
\alias{plot_optimization_search}
\title{Plot Search Space Exploration}
\usage{
plot_optimization_search(result, color_by = "eval", ...)
}
\arguments{
\item{result}{rctbp_optimization_result object}

\item{color_by}{Column to color points by (default: evaluation order)}

\item{...}{Additional ggplot arguments}
}
\value{
ggplot2 object
}
\description{
Visualizes how the optimizer explored the search space.
}
