% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimization.R
\name{plot_pareto_convergence}
\alias{plot_pareto_convergence}
\title{Plot Convergence for Pareto Optimization}
\usage{
plot_pareto_convergence(result, show_all = TRUE, ...)
}
\arguments{
\item{result}{rctbp_pareto_result object}

\item{show_all}{Logical, show all evaluations}

\item{...}{Additional ggplot arguments}
}
\value{
ggplot2 object
}
\description{
Plot Convergence for Pareto Optimization
}
\keyword{internal}
