% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimization.R
\name{plot_pareto_search}
\alias{plot_pareto_search}
\title{Plot Search Space for Pareto Optimization}
\usage{
plot_pareto_search(result, ...)
}
\arguments{
\item{result}{rctbp_pareto_result object}

\item{...}{Additional ggplot arguments}
}
\value{
ggplot2 object
}
\description{
Plot Search Space for Pareto Optimization
}
\keyword{internal}
