% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{print.rctbayespower}
\alias{print.rctbayespower}
\title{Print method for rctbayespower objects}
\usage{
\method{print}{rctbayespower}(x, ...)
}
\arguments{
\item{x}{An rctbayespower object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for rctbayespower objects
}
