% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{print.rctbayespower_curve}
\alias{print.rctbayespower_curve}
\title{Print method for power curve objects}
\usage{
\method{print}{rctbayespower_curve}(x, ...)
}
\arguments{
\item{x}{An rctbayespower_curve object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for power curve objects
}
