% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_analysis.R
\name{print.rctbayespower_effectsize}
\alias{print.rctbayespower_effectsize}
\title{Print method for effect size analysis objects}
\usage{
\method{print}{rctbayespower_effectsize}(x, ...)
}
\arguments{
\item{x}{An rctbayespower_effectsize object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for effect size analysis objects
}
