% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{print.rctbayespower_samplesize}
\alias{print.rctbayespower_samplesize}
\title{Print method for sample size analysis objects}
\usage{
\method{print}{rctbayespower_samplesize}(x, ...)
}
\arguments{
\item{x}{An rctbayespower_samplesize object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for sample size analysis objects
}
