% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{print.rctbp_model}
\alias{print.rctbp_model}
\title{Print Method for rctbp_model Objects}
\arguments{
\item{x}{An S7 object of class "rctbp_model"}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for rctbp_model Objects
}
