% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_objectives.R
\name{print.rctbp_objectives}
\alias{print.rctbp_objectives}
\title{Print Method for rctbp_objectives}
\arguments{
\item{x}{An rctbp_objectives object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns x
}
\description{
Print Method for rctbp_objectives
}
