% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_objectives.R
\name{print.rctbp_optimization_result}
\alias{print.rctbp_optimization_result}
\title{Print Method for rctbp_optimization_result}
\arguments{
\item{x}{An rctbp_optimization_result object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns x
}
\description{
Print Method for rctbp_optimization_result
}
