% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sample_size_result.R
\name{print.rctbp_sample_size_result}
\alias{print.rctbp_sample_size_result}
\title{Print Method for rctbp_sample_size_result}
\arguments{
\item{x}{An rctbp_sample_size_result object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns x
}
\description{
Print Method for rctbp_sample_size_result
}
