% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{replace_arg_with_symbol}
\alias{replace_arg_with_symbol}
\title{Replace a Call Argument with a Symbol}
\usage{
replace_arg_with_symbol(call_obj, arg_name, symbol_name)
}
\arguments{
\item{call_obj}{A call object}

\item{arg_name}{The argument name to replace}

\item{symbol_name}{The symbol name to substitute}
}
\value{
Modified call object
}
\description{
Replaces a named argument in a call object with a symbol reference
(e.g., replace \verb{design = <actual object>} with \code{design = design}).
}
\keyword{internal}
