% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_grid_analysis.R
\name{resolve_interim_parameters}
\alias{resolve_interim_parameters}
\title{Resolve Interim Parameters from Condition}
\usage{
resolve_interim_parameters(condition, design, static_overrides = list())
}
\arguments{
\item{condition}{List containing condition-specific parameters}

\item{design}{rctbayespower_design object}

\item{static_overrides}{List of static parameter overrides}
}
\value{
Named list of resolved interim parameters or NULL if no interim function
}
\description{
Maps condition values to interim function parameters when interim analyses
are specified in the condition.
}
\keyword{internal}
