% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_single.R
\name{run_gp_power_loop}
\alias{run_gp_power_loop}
\title{Run GP Power Custom Loop}
\usage{
run_gp_power_loop(
  design,
  target_power,
  n_range,
  constant,
  n_sims,
  n_cores,
  score_shape,
  score_scale,
  n_init,
  max_evals,
  patience,
  verbose,
  bf_args,
  brms_args
)
}
\description{
Fits GP to logit(power) on log(n) scale. Uses P_feas acquisition rule
to select next n. Does not use bbotk.
}
\keyword{internal}
