% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{sample_from_npe_posterior}
\alias{sample_from_npe_posterior}
\title{Sample from NPE Posterior Distribution}
\usage{
sample_from_npe_posterior(posterior_params, n_samples, batch_size)
}
\arguments{
\item{posterior_params}{Output from NPE model forward pass. Format depends
on architecture (e.g., mean/variance for Gaussian, flow parameters, etc.)}

\item{n_samples}{Number of posterior samples to generate per simulation}

\item{batch_size}{Number of simulations in the batch}
}
\value{
Array of shape \verb{[batch_size, n_samples, n_params]}
}
\description{
Generates posterior samples from the parameters learned by the NPE model.
The sampling strategy depends on the type of NPE architecture used.
}
\keyword{internal}
