% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_analysis.R
\name{sample_size_analysis}
\alias{sample_size_analysis}
\title{Sample Size Analysis for Bayesian RCTs}
\usage{
sample_size_analysis(
  effect_size,
  target_power = 0.8,
  outcome_type = "continuous",
  sample_sizes = seq(20, 200, by = 10),
  allocation_ratio = 1,
  n_simulations = 500,
  ...
)
}
\arguments{
\item{effect_size}{Expected effect size}

\item{target_power}{Desired power level (default: 0.8)}

\item{outcome_type}{Type of outcome ("continuous", "binary", "count")}

\item{sample_sizes}{Vector of sample sizes to test (per group)}

\item{allocation_ratio}{Ratio of treatment to control (default: 1)}

\item{n_simulations}{Number of simulations per sample size}

\item{...}{Additional arguments passed to power_analysis()}
}
\value{
A list containing sample size analysis results
}
\description{
Determine optimal sample size for achieving desired power in Bayesian RCTs.
This function runs power analyses across a range of sample sizes to find
the minimum sample size needed to achieve target power.
}
\examples{
\dontrun{
# Find sample size for 80\% power
sample_size_result <- sample_size_analysis(
  effect_size = 0.5,
  target_power = 0.8,
  outcome_type = "continuous",
  sample_sizes = seq(20, 100, by = 10),
  n_simulations = 100
)
}
}
