% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_optimize.R
\name{setup_pareto_mbo}
\alias{setup_pareto_mbo}
\title{Setup MBO Components for Pareto Optimization}
\usage{
setup_pareto_mbo(instance, surrogate, verbosity)
}
\arguments{
\item{instance}{bbotk instance}

\item{surrogate}{Surrogate type}

\item{verbosity}{Verbosity level}
}
\value{
List with surrogate, acq_function, acq_optimizer, optimizer
}
\description{
Setup MBO Components for Pareto Optimization
}
\keyword{internal}
