% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity.R
\name{should_show}
\alias{should_show}
\title{Check if Message Should Be Displayed}
\usage{
should_show(required_level = 1)
}
\arguments{
\item{required_level}{Integer: minimum verbosity level required to show message
\itemize{
\item 0 = always show (errors, critical messages)
\item 1 = show in normal and verbose modes (standard messages)
\item 2 = show only in verbose mode (debug messages)
}}
}
\value{
Logical: TRUE if message should be displayed, FALSE otherwise
}
\description{
Determines whether a message at a given level should be shown
based on current verbosity settings.
}
\examples{
# Check if verbose messages should be shown
if (should_show(2)) {
  message("This is a debug message")
}

}
