% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{show_optimization_args}
\alias{show_optimization_args}
\title{Show Available Optimization Arguments}
\usage{
show_optimization_args()
}
\value{
Invisibly returns NULL
}
\description{
Displays the available arguments and their defaults for \code{\link[=optimization]{optimization()}}.
}
\examples{
show_optimization_args()
}
