% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pareto_result.R, R/s3_wrappers.R
\name{summary.rctbp_pareto_result}
\alias{summary.rctbp_pareto_result}
\title{Summary Method for rctbp_pareto_result}
\usage{
\method{summary}{rctbp_pareto_result}(object, ...)
}
\arguments{
\item{object}{An rctbp_pareto_result object}

\item{...}{Additional arguments (unused)}
}
\value{
The Pareto front data frame
}
\description{
Summary Method for rctbp_pareto_result
}
