% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sample_size_result.R
\name{summary.rctbp_sample_size_result}
\alias{summary.rctbp_sample_size_result}
\title{Summary Method for rctbp_sample_size_result}
\arguments{
\item{object}{An rctbp_sample_size_result object}

\item{...}{Additional arguments (unused)}
}
\value{
The archive data frame
}
\description{
Summary Method for rctbp_sample_size_result
}
