% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_optimize.R
\name{suppress_mbo_logging}
\alias{suppress_mbo_logging}
\title{Suppress MBO Logging}
\usage{
suppress_mbo_logging(verbosity)
}
\arguments{
\item{verbosity}{Verbosity level}
}
\description{
Suppress MBO Logging
}
\keyword{internal}
