% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_internal.R
\name{unstandardize_log_n}
\alias{unstandardize_log_n}
\title{Unstandardize Log-n}
\usage{
unstandardize_log_n(z, mean_log, sd_log)
}
\arguments{
\item{z}{Standardized value}

\item{mean_log}{Mean used for standardization}

\item{sd_log}{SD used for standardization}
}
\value{
Sample size on original scale
}
\description{
Converts standardized log(n) back to original n scale.
}
\keyword{internal}
