% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity.R
\name{with_verbosity}
\alias{with_verbosity}
\title{Temporarily Change Verbosity Level}
\usage{
with_verbosity(verbosity = c(0, 1, 2), code)
}
\arguments{
\item{verbosity}{Integer: verbosity level (0, 1, or 2)}

\item{code}{Code to execute with the temporary verbosity}
}
\value{
Result of evaluating \code{code}
}
\description{
Executes code with a temporary verbosity level.
}
\examples{
\dontrun{
# Run analysis in quiet mode temporarily
with_verbosity(0, {
  result <- power_analysis(conditions, n_sims = 100)
})
}

}
