% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq_function_eic.R
\name{AcqFunctionEIC}
\alias{AcqFunctionEIC}
\title{Expected Improvement with Constraints (EIC) Acquisition Function}
\description{
Custom acquisition function that multiplies Expected Improvement by the
probability of feasibility. Optimized for sample size optimization where:
\itemize{
\item Objective (n_total) is an input parameter, giving deterministic EI
\item Constraint (power) is an output, modeled with GP on logit scale
}
}
\details{
The acquisition value is: \code{EIC(x) = EI(x) * P(power(x) >= target)}

For sample size optimization (objective is input):
\itemize{
\item \code{EI(x) = max(0, n_best - n)} where n_best is smallest feasible n observed
}

P(feasible) is computed from the constraint surrogate on logit scale:
\code{P(power >= target) = Phi((mu_logit - logit(target)) / sigma_logit)}

Optional conservatism via \code{kappa} parameter uses lower confidence bound:
\code{effective_mean = mu_logit - kappa * sigma_logit}
}
\section{Kappa values}{

\itemize{
\item \code{kappa = 0}: Uses posterior mean for P(feasible). Neutral, may recommend
sample sizes with ~50\% chance of meeting target.
\item \code{kappa = 1.96}: Uses 97.5\% lower confidence bound. Recommendations have
high probability (97.5\%) of meeting the target.
\item \code{kappa = 2.58}: Uses 99.5\% LCB. Very conservative.
}
}

\section{Fields}{

\describe{
\item{surrogate_constraint}{Surrogate for constraint (power on logit scale)}
\item{target_value}{Target value on original scale (e.g., 0.80 for power)}
\item{target_logit}{Target on logit scale}
\item{kappa}{Conservatism parameter (0 = neutral, 1.96 = 97.5\% LCB)}
\item{y_best}{Current best feasible objective value (for EI computation)}
\item{constraint_col}{Name of the constraint column in archive}
\item{objective_col}{Name of the objective column (input for minimization)}
\item{objective_is_input}{TRUE if objective is an input (e.g., n_total)}
}
}

\section{Methods}{

\describe{
\item{\code{initialize(surrogate_objective, surrogate_constraint, target_value, kappa, constraint_col, objective_col, objective_is_input)}}{Create EIC acquisition function}
\item{\code{update(archive)}}{Update with current archive}
}
}

\section{Super classes}{
\code{\link[bbotk:Objective]{bbotk::Objective}} -> \code{\link[mlr3mbo:AcqFunction]{mlr3mbo::AcqFunction}} -> \code{AcqFunctionEIC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AcqFunctionEIC-new}{\code{AcqFunctionEIC$new()}}
\item \href{#method-AcqFunctionEIC-update}{\code{AcqFunctionEIC$update()}}
\item \href{#method-AcqFunctionEIC-.eval_dt}{\code{AcqFunctionEIC$.eval_dt()}}
\item \href{#method-AcqFunctionEIC-clone}{\code{AcqFunctionEIC$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="eval"><a href='../../bbotk/html/Objective.html#method-Objective-eval'><code>bbotk::Objective$eval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="format"><a href='../../bbotk/html/Objective.html#method-Objective-format'><code>bbotk::Objective$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="help"><a href='../../bbotk/html/Objective.html#method-Objective-help'><code>bbotk::Objective$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="print"><a href='../../bbotk/html/Objective.html#method-Objective-print'><code>bbotk::Objective$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqFunction" data-id="eval_dt"><a href='../../mlr3mbo/html/AcqFunction.html#method-AcqFunction-eval_dt'><code>mlr3mbo::AcqFunction$eval_dt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqFunction" data-id="eval_many"><a href='../../mlr3mbo/html/AcqFunction.html#method-AcqFunction-eval_many'><code>mlr3mbo::AcqFunction$eval_many()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqFunction" data-id="reset"><a href='../../mlr3mbo/html/AcqFunction.html#method-AcqFunction-reset'><code>mlr3mbo::AcqFunction$reset()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcqFunctionEIC-new"></a>}}
\if{latex}{\out{\hypertarget{method-AcqFunctionEIC-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionEIC$new(
  surrogate_objective,
  surrogate_constraint,
  target_value = 0.8,
  kappa = 0,
  constraint_col = "logit_pwr_eff",
  objective_col = "n_total",
  objective_is_input = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcqFunctionEIC-update"></a>}}
\if{latex}{\out{\hypertarget{method-AcqFunctionEIC-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionEIC$update(archive)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcqFunctionEIC-.eval_dt"></a>}}
\if{latex}{\out{\hypertarget{method-AcqFunctionEIC-.eval_dt}{}}}
\subsection{Method \code{.eval_dt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionEIC$.eval_dt(xdt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcqFunctionEIC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AcqFunctionEIC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionEIC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
