% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSE.R
\name{calculate_mcse_mean}
\alias{calculate_mcse_mean}
\title{Monte Carlo Standard Error for Continuous Metrics}
\usage{
calculate_mcse_mean(values, n_sims)
}
\arguments{
\item{values}{Vector of continuous values}

\item{n_sims}{Total number of simulations}
}
\value{
Monte Carlo Standard Error
}
\description{
Calculate Monte Carlo Standard Error for continuous metrics like mean probabilities.
}
\keyword{internal}
