% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design.R
\name{create_ancova_sim_fn}
\alias{create_ancova_sim_fn}
\title{Create ANCOVA Simulation Function}
\usage{
create_ancova_sim_fn(n_arms)
}
\arguments{
\item{n_arms}{Number of arms}
}
\value{
Simulation function
}
\description{
Internal helper to create simulation function for ANCOVA models.
}
\keyword{internal}
