% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_optimize.R
\name{create_pareto_objective_fn}
\alias{create_pareto_objective_fn}
\title{Create Objective Function for Pareto Optimization}
\usage{
create_pareto_objective_fn(
  design,
  objectives,
  search,
  search_specs,
  constant,
  constraint,
  n_sims,
  n_cores,
  bf_args,
  brms_args,
  verbosity
)
}
\arguments{
\item{design}{rctbp_design object}

\item{objectives}{Named list of objectives}

\item{search}{Named list of search bounds}

\item{search_specs}{Parsed simplex specs}

\item{constant}{Fixed parameters}

\item{constraint}{Optional constraint}

\item{n_sims}{Number of simulations per evaluation}

\item{n_cores}{Number of cores}

\item{bf_args}{BayesFlow arguments}

\item{brms_args}{brms arguments}

\item{verbosity}{Verbosity level}
}
\value{
List with fn (objective function)
}
\description{
Create Objective Function for Pareto Optimization
}
\keyword{internal}
