% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_power_analysis.R
\name{power_analysis}
\alias{power_analysis}
\title{Execute Power Analysis}
\usage{
power_analysis(object, ...)
}
\arguments{
\item{object}{An S7 object to execute (must be rctbp_power_analysis)}

\item{...}{Additional arguments passed to specific methods}
}
\value{
The modified rctbp_power_analysis object with results stored in the
\code{sim_results} slot.
}
\description{
S7 generic function for executing power analysis objects. This function provides
the interface for running power analysis configurations created with the
power_analysis() constructor.
}
\details{
This S7 generic function dispatches to the appropriate method based on the class
of the input object. Currently supported:
\itemize{
\item \code{rctbp_power_analysis}: Executes Bayesian power analysis
}
}
\examples{
\dontrun{
# Create and execute power analysis
power_config <- power_analysis(conditions, n_sims = 100)
power_config <- power_analysis(power_config)
}
}
\seealso{
\code{\link[=power_analysis]{power_analysis()}} for creating power analysis configurations
}
