% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{print.rctbayespower}
\alias{print.rctbayespower}
\title{Print method for rctbayespower objects}
\usage{
\method{print}{rctbayespower}(x, ...)
}
\arguments{
\item{x}{An object of class "rctbayespower" returned by power_analysis()}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object. Used for side effects (printing).
}
\description{
Print method for rctbayespower objects
}
