% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctbayespower-package.R
\docType{package}
\name{rctbayespower-package}
\alias{rctbayespower-package}
\alias{rctbayespower}
\title{The 'rctbayespower' package.}
\description{
Bayesian power analysis for randomized controlled trials (RCTs) using brms and Stan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matthiaskloft.github.io/rctbayespower/}
  \item \url{https://github.com/matthiaskloft/rctbayespower}
}

}
\author{
\strong{Maintainer}: Matthias Kloft \email{kloft.dev+rctbayespower@gmail.com} (\href{https://orcid.org/0000-0003-1845-6957}{ORCID}) [copyright holder]

}
\keyword{internal}
