% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_single_run.R
\name{simulate_single_run}
\alias{simulate_single_run}
\title{Single Run Simulation for RCT Bayesian Power Analysis (DEPRECATED)}
\usage{
simulate_single_run(condition_arguments, id_iter, design)
}
\arguments{
\item{condition_arguments}{A single entry from the condition_arguments list}

\item{id_iter}{Iteration identifier}

\item{design}{A rctbp_design object}
}
\value{
A data frame with power analysis measures
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Use \code{\link[=worker_process_single]{worker_process_single()}} instead.
}
\keyword{internal}
