% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_grid_analysis.R
\name{summary.rctbayespower_grid}
\alias{summary.rctbayespower_grid}
\title{Summary method for power grid analysis objects}
\usage{
\method{summary}{rctbayespower_grid}(object, ...)
}
\arguments{
\item{object}{An rctbayespower_grid object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for power grid analysis objects
}
